backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bs4>=0.0.2",
    "chromadb>=1.4.1",
    "python-dotenv>=1.0.0",
    "fastapi[standard]>=0.128.2",
    "llama-index-core>=0.14.13",
    "llama-index-embeddings-openai-like>=0.2.2",
    "llama-index-llms-openai-like>=0.6.0",
    "llama-index-vector-stores-chroma>=0.5.5",
    "markdownify>=1.2.2",
    "mcp>=1.26.0",
]
